Move Talon class and MotorController State to org.usfirst.frc2013.Robot2018.motor package.  Trying to move motor library stuff out of the basic robot framework package.
