+++ Robot2018/src/org/usfirst/frc2813/Robot2018/commands/Elevator/ElevatorHoldPosition.java
@@ -3,7 +3,7 @@ package org.usfirst.frc2813.Robot2018.commands.Elevator;
 import org.usfirst.frc2813.Robot2018.Robot;
 import org.usfirst.frc2813.Robot2018.commands.GearheadsCommand;
 import org.usfirst.frc2813.Robot2018.commands.GearheadsInstantCommand;
-import org.usfirst.frc2813.Robot2018.MotorControllerState;
+import org.usfirst.frc2813.Robot2018.motor.MotorControllerState;
 
 /**
  * Maintain elevator position. Elevator does this actively with
