+++ Robot2018/src/org/usfirst/frc2813/Robot2018/motor/Talon.java
@@ -1,8 +1,7 @@
-package org.usfirst.frc2813.Robot2018;
+package org.usfirst.frc2813.Robot2018.motor;
 
 import java.util.logging.Logger;
 
-import org.usfirst.frc2813.Robot2018.motor.TalonPID;
 import org.usfirst.frc2813.util.unit.Direction;
 
 import com.ctre.phoenix.motorcontrol.can.TalonSRX;
