+++ Robot2018/src/org/usfirst/frc2813/Robot2018/subsystems/Elevator.java
@@ -2,9 +2,9 @@ package org.usfirst.frc2813.Robot2018.subsystems;
 
 
 import org.usfirst.frc2813.Robot2018.RobotMap;
-import org.usfirst.frc2813.Robot2018.Talon;
-import org.usfirst.frc2813.Robot2018.MotorControllerState;
 import org.usfirst.frc2813.Robot2018.commands.Elevator.ElevatorHoldPosition;
+import org.usfirst.frc2813.Robot2018.motor.MotorControllerState;
+import org.usfirst.frc2813.Robot2018.motor.Talon;
 import org.usfirst.frc2813.Robot2018.motor.TalonSensorPhase;
 import org.usfirst.frc2813.util.unit.Direction;
